requirements.txt
langchain==0.3.15
langgraph==0.2.63
langchain-google-genai==2.0.8
google-generativeai==0.8.3
python-dotenv==1.0.1
